+++ packages/core/src/interpreter.ts
@@ -59,13 +59,7 @@ import { isInFinalState } from './stateUtils';
 import { registry } from './registry';
 import { getGlobal, registerService } from './devTools';
 import * as serviceScope from './serviceScope';
-import {
-  ActorRef,
-  ActorRefFrom,
-  SpawnedActorRef,
-  StopActionObject,
-  Subscription
-} from '.';
+import { ActorRef, ActorRefFrom, StopActionObject, Subscription } from '.';
 
 export type StateListener<
   TContext,
@@ -113,7 +107,8 @@ export class Interpreter<
   TStateSchema extends StateSchema = any,
   TEvent extends EventObject = EventObject,
   TTypestate extends Typestate<TContext> = { value: any; context: TContext }
-> implements Actor<State<TContext, TEvent, TStateSchema, TTypestate>, TEvent> {
+> implements
+    ActorRef<TEvent, State<TContext, TEvent, TStateSchema, TTypestate>> {
   /**
    * The default interpreter options:
    *
@@ -170,7 +165,7 @@ export class Interpreter<
    * The globally unique process ID for this invocation.
    */
   public sessionId: string;
-  public children: Map<string | number, SpawnedActorRef<any>> = new Map();
+  public children: Map<string | number, ActorRef<any>> = new Map();
   private forwardTo: Set<string> = new Set();
 
   // Dev Tools
@@ -945,7 +940,7 @@ export class Interpreter<
     entity: Spawnable,
     name: string,
     options?: SpawnOptions
-  ): SpawnedActorRef<any> {
+  ): ActorRef<any> {
     if (isPromiseLike(entity)) {
       return this.spawnPromise(Promise.resolve(entity), name);
     } else if (isFunction(entity)) {
@@ -969,7 +964,7 @@ export class Interpreter<
   >(
     machine: StateMachine<TChildContext, TChildStateSchema, TChildEvent>,
     options: { id?: string; autoForward?: boolean; sync?: boolean } = {}
-  ): SpawnedActorRef<TChildEvent, State<TChildContext, TChildEvent>> {
+  ): ActorRef<TChildEvent, State<TChildContext, TChildEvent>> {
     const childService = new Interpreter(machine, {
       ...this.options, // inherit options from this interpreter
       parent: this,
@@ -1007,15 +1002,14 @@ export class Interpreter<
 
     return actor;
   }
-  private spawnPromise<T>(
-    promise: Promise<T>,
-    id: string
-  ): SpawnedActorRef<never, T> {
+  private spawnPromise<T>(promise: Promise<T>, id: string): ActorRef<never, T> {
     let canceled = false;
+    let resolvedData: T | undefined = undefined;
 
     promise.then(
       (response) => {
         if (!canceled) {
+          resolvedData = response;
           this.removeChild(id);
           this.send(
             toSCXMLEvent(doneInvoke(id, response) as any, { origin: id })
@@ -1046,7 +1040,7 @@ export class Interpreter<
       }
     );
 
-    const actor: SpawnedActorRef<never, T> = {
+    const actor: ActorRef<never, T> = {
       id,
       send: () => void 0,
       subscribe: (next, handleError?, complete?) => {
@@ -1081,22 +1075,22 @@ export class Interpreter<
       },
       toJSON() {
         return { id };
-      }
+      },
+      getSnapshot: () => resolvedData
     };
 
     this.children.set(id, actor);
 
     return actor;
   }
-  private spawnCallback(
-    callback: InvokeCallback,
-    id: string
-  ): SpawnedActorRef<any> {
+  private spawnCallback(callback: InvokeCallback, id: string): ActorRef<any> {
     let canceled = false;
     const receivers = new Set<(e: EventObject) => void>();
     const listeners = new Set<(e: EventObject) => void>();
+    let emitted: TEvent | undefined = undefined;
 
     const receive = (e: TEvent) => {
+      emitted = e;
       listeners.forEach((listener) => listener(e));
       if (canceled) {
         return;
@@ -1140,7 +1134,8 @@ export class Interpreter<
       },
       toJSON() {
         return { id };
-      }
+      },
+      getSnapshot: () => emitted
     };
 
     this.children.set(id, actor);
@@ -1150,9 +1145,12 @@ export class Interpreter<
   private spawnObservable<T extends TEvent>(
     source: Subscribable<T>,
     id: string
-  ): SpawnedActorRef<any, T> {
+  ): ActorRef<any, T> {
+    let emitted: T | undefined = undefined;
+
     const subscription = source.subscribe(
       (value) => {
+        emitted = value;
         this.send(toSCXMLEvent(value, { origin: id }));
       },
       (err) => {
@@ -1165,13 +1163,14 @@ export class Interpreter<
       }
     );
 
-    const actor: SpawnedActorRef<any, T> = {
+    const actor: ActorRef<any, T> = {
       id,
       send: () => void 0,
       subscribe: (next, handleError?, complete?) => {
         return source.subscribe(next, handleError, complete);
       },
       stop: () => subscription.unsubscribe(),
+      getSnapshot: () => emitted,
       toJSON() {
         return { id };
       }
@@ -1181,7 +1180,7 @@ export class Interpreter<
 
     return actor;
   }
-  private spawnActor<T extends SpawnedActorRef<any>>(actor: T): T {
+  private spawnActor<T extends ActorRef<any>>(actor: T): T {
     this.children.set(actor.id, actor);
 
     return actor;
@@ -1215,6 +1214,7 @@ export class Interpreter<
         return { unsubscribe: () => void 0 };
       },
       stop: dispose || undefined,
+      getSnapshot: () => undefined,
       toJSON() {
         return { id };
       }
@@ -1267,6 +1267,13 @@ export class Interpreter<
   public [symbolObservable]() {
     return this;
   }
+
+  public getSnapshot() {
+    if (this.status === InterpreterStatus.NotStarted) {
+      return this.initialState;
+    }
+    return this._state!;
+  }
 }
 
 const resolveSpawnOptions = (nameOrOptions?: string | SpawnOptions) => {
@@ -1288,11 +1295,11 @@ export function spawn<TC, TE extends EventObject>(
 export function spawn(
   entity: Spawnable,
   nameOrOptions?: string | SpawnOptions
-): SpawnedActorRef<any>;
+): ActorRef<any>;
 export function spawn(
   entity: Spawnable,
   nameOrOptions?: string | SpawnOptions
-): SpawnedActorRef<any> {
+): ActorRef<any> {
   const resolvedOptions = resolveSpawnOptions(nameOrOptions);
 
   return serviceScope.consume((service) => {
@@ -1305,6 +1312,7 @@ export function spawn(
         }") outside of a service. This will have no effect.`
       );
     }
+
     if (service) {
       return service.spawn(entity, resolvedOptions.name, resolvedOptions);
     } else {