+++ packages/xstate-react/test/useActor.test.tsx
@@ -6,9 +6,11 @@ import {
   assign,
   spawn,
   ActorRef,
-  ActorRefFrom
+  ActorRefFrom,
+  interpret
 } from 'xstate';
-import { render, cleanup, fireEvent } from '@testing-library/react';
+import { toActorRef } from 'xstate/lib/Actor';
+import { render, cleanup, fireEvent, act } from '@testing-library/react';
 import { useActor } from '../src/useActor';
 import { useState } from 'react';
 
@@ -241,7 +243,7 @@ describe('useActor', () => {
   });
 
   it('actor should provide snapshot value immediately', () => {
-    const simpleActor: ActorRef<any, number> & { latestValue: number } = {
+    const simpleActor = toActorRef({
       send: () => {
         /* ... */
       },
@@ -253,6 +255,8 @@ describe('useActor', () => {
           }
         };
       }
+    }) as ActorRef<any, number> & {
+      latestValue: number;
     };
 
     const Test = () => {
@@ -268,14 +272,13 @@ describe('useActor', () => {
     expect(div.textContent).toEqual('42');
   });
 
-  it('should update snapshot value when actor changes', () => {
-    const createSimpleActor = (
-      value: number
-    ): ActorRef<any, number> & { latestValue: number } => ({
+  it('should provide value from `actor.getSnapshot()`', () => {
+    const simpleActor = toActorRef({
+      id: 'test',
       send: () => {
         /* ... */
       },
-      latestValue: value,
+      getSnapshot: () => 42,
       subscribe: () => {
         return {
           unsubscribe: () => {
@@ -284,6 +287,35 @@ describe('useActor', () => {
         };
       }
     });
+
+    const Test = () => {
+      const [state] = useActor(simpleActor);
+
+      return <div data-testid="state">{state}</div>;
+    };
+
+    const { getByTestId } = render(<Test />);
+
+    const div = getByTestId('state');
+
+    expect(div.textContent).toEqual('42');
+  });
+
+  it('should update snapshot value when actor changes', () => {
+    const createSimpleActor = (value: number) =>
+      toActorRef({
+        send: () => {
+          /* ... */
+        },
+        latestValue: value,
+        subscribe: () => {
+          return {
+            unsubscribe: () => {
+              /* ... */
+            }
+          };
+        }
+      }) as ActorRef<any> & { latestValue: number };
 
     const Test = () => {
       const [actor, setActor] = useState(createSimpleActor(42));
@@ -322,16 +354,16 @@ describe('useActor', () => {
     const noop = () => {
       /* ... */
     };
-    const firstActor: ActorRef<any> = {
+    const firstActor = toActorRef({
       send: noop,
       subscribe: fakeSubscribe
-    };
-    const lastActor: ActorRef<any> = {
+    });
+    const lastActor = toActorRef({
       send: () => {
         done();
       },
       subscribe: fakeSubscribe
-    };
+    });
 
     const Test = () => {
       const [actor, setActor] = useState(firstActor);
@@ -368,5 +400,74 @@ describe('useActor', () => {
     // The effect will call the closed-in `send`, which originally
     // was the reference to the first actor. Now that `send` is stable,
     // it will always refer to the latest actor.
+  });
+
+  it('should also work with services', () => {
+    const counterMachine = createMachine<
+      { count: number },
+      { type: 'INC' } | { type: 'SOMETHING' }
+    >(
+      {
+        id: 'counter',
+        initial: 'active',
+        context: { count: 0 },
+        states: {
+          active: {
+            on: {
+              INC: { actions: assign({ count: (ctx) => ctx.count + 1 }) },
+              SOMETHING: { actions: 'doSomething' }
+            }
+          }
+        }
+      },
+      {
+        actions: {
+          doSomething: () => {
+            /* do nothing */
+          }
+        }
+      }
+    );
+    const counterService = interpret(counterMachine).start();
+
+    const Counter = () => {
+      const [state, send] = useActor(counterService);
+
+      return (
+        <div
+          data-testid="count"
+          onClick={() => {
+            send('INC');
+            // @ts-expect-error
+            send('FAKE');
+          }}
+        >
+          {state.context.count}
+        </div>
+      );
+    };
+
+    const { getAllByTestId } = render(
+      <>
+        <Counter />
+        <Counter />
+      </>
+    );
+
+    const countEls = getAllByTestId('count');
+
+    expect(countEls.length).toBe(2);
+
+    countEls.forEach((countEl) => {
+      expect(countEl.textContent).toBe('0');
+    });
+
+    act(() => {
+      counterService.send({ type: 'INC' });
+    });
+
+    countEls.forEach((countEl) => {
+      expect(countEl.textContent).toBe('1');
+    });
   });
 });